+++ structure.py
@@ -388,7 +388,7 @@ class Positions(Dat):
 
         Returns
         -------
-        time0 : (*,) float numpy array
+        time0 : (*,) int numpy array
             Array of frames.
         """
