+++ work.py
@@ -726,7 +726,7 @@ class ActiveWork(Dat):
 
         Returns
         -------
-        time0 : (*,) float numpy array
+        time0 : (*,) int numpy array
             Array of initial times.
         """
